requirements.txt
# Data processing
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
matplotlib

# Machine Learning
scikit-learn>=1.2.0
scipy>=1.10.0

# API framework
fastapi>=0.95.0
uvicorn>=0.20.0

# Database
sqlalchemy>=2.0.0
# sqlite3  # Built into Python

# Utilities
python-dotenv>=1.0.0
pydantic>=1.10.0

# Development
jupyter>=1.0.0
pytest>=7.0.0